requirements.txt
docker
python-dotenv
groq
